Custom loading screen image
Not sure if I'm doing this right but hey might as well at least try something. This is to add a custom image to the initial loading screen when you first join the server. Just thought it would add another unique thing to it. You can view the image I put together here:

http://postimg.org/image/9xnkhvnx7/

Credit for this code from Wiking @ Arma Junkies.